backend/pyproject.toml
[project]
name = "FarmAtlas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pandera[pandas]>=0.26.1",
    "plotly>=6.4.0",
    "pyet>=1.4.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.41.0",
]
